refactor(oci): share canonical ocx_cli_identifier helper
Introduce public oci::ocx_cli_identifier() returning the canonical
untagged ocx/cli identifier, and fold the __OCX_SELF_IMAGE/__testing
loopback-redirect override into that single helper. Migrate the four
inline new_registry("ocx/cli", ...) reconstruction sites (activate
ocx_install_bin_path, update_check self_update error path + the two
callers of the former module-private helper) to consume it.

Behavior-preserving (Two Hats: refactor only); seam unit tests moved
alongside the helper. Precursor to the self-setup bootstrap.

Refs #148
